+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Components/Lyrics/Carets/DrawableTimeTagRecordCursor.cs
@@ -7,11 +7,12 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Graphics.Shapes;
+using osu.Game.Rulesets.Karaoke.Objects;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Components
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
 {
-    public class DrawableTimeTagEditCursor : CompositeDrawable, IDrawableCursor, IHasCursorPosition
+    public class DrawableTimeTagRecordCursor : CompositeDrawable, IDrawableCursor, IHasTimeTag
     {
         private const float triangle_width = 8;
 
@@ -20,7 +21,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Components
 
         private readonly RightTriangle drawableTimeTag;
 
-        public DrawableTimeTagEditCursor()
+        public DrawableTimeTagRecordCursor()
         {
             AutoSizeAxes = Axes.Both;
 
@@ -33,18 +34,16 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Components
             };
         }
 
-        private CursorPosition position;
+        private TimeTag timeTag;
 
-        public CursorPosition CursorPosition
+        public TimeTag TimeTag
         {
-            get => position;
+            get => timeTag;
             set
             {
-                position = value;
-                drawableTimeTag.Scale = new Vector2(position.Index.State == TextIndex.IndexState.Start ? 1 : -1, 1);
-
-                // todo : color is by has time-tag here?
-                // drawableTimeTag.Colour = position.Time.HasValue ? colours.YellowDarker : colours.Gray3;
+                timeTag = value;
+                drawableTimeTag.Scale = new Vector2(timeTag.Index.State == TextIndex.IndexState.Start ? 1 : -1, 1);
+                drawableTimeTag.Colour = timeTag.Time.HasValue ? colours.YellowDarker : colours.Gray3;
             }
         }
