+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Components/Lyrics/Carets/IHasTimeTag.cs
@@ -3,7 +3,7 @@
 
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Components
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
 {
     public interface IHasTimeTag
     {
