+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Components/Lyrics/LyricControl.cs
@@ -8,7 +8,8 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input.Events;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Components;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Parts;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
 using osu.Game.Screens.Edit;
